+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.1.0]
+
+* **\[BREAKING CHANGES\]**: Whole code has been rewritten, and no longer utilizing callback style. Please provide a Future into the _computation_ argument.
+* **\[BREAKING CHANGES\]**: Methods are accessible using static accesses.
+* New static method _**setDefaultWidgetBuilder**_: Provides a function to define the Widget to be displayed during the loading process.
+* Updated **README.md**
+* Updated _**main.dart**_ example.
+ 
+
 ## [0.0.3]
 
 * Added default WidgetBuilder option